Feature/acess control refactor

+++ src/main/java/com/jpoltramari/library_api/domain/repository/GroupRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface GroupRepository extends CustomJpaRepository<Group, Short> {
+public interface GroupRepository extends CustomJpaRepository<Group, Long> {
 
     Optional<Group> findByName(String name);
 }
